Add gdcc/nginx:unstable image and dev_nginx container
Added a nginx container to the docker setup. This could be used for
various proxying stuff, but it is primarily added to work around a problem
with the IntelliJ Payara plugin. It doesn't allow remote redeployment in
case the Payara admin is served via HTTPS using a self-signed certificate,
which is the case of the default dataverse container installation.

This nginx configuration provides an HTTP endpoint at port 4849, and proxies
requests to the Payara admin console's HTTPS 4848 endpoint.

From the IntelliJ Payara plugin one has to specify the localhost 4849 port
(without SSL).